+++ src/main/java/ru/practicum/shareit/user/storage/UserStorage.java
@@ -5,7 +5,7 @@ import ru.practicum.shareit.user.model.User;
 
 import java.util.Optional;
 
-public interface UserStorage extends JpaRepository<User, Integer> {
+public interface UserStorage extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String url);
